Pin build-pypi-wheel runner image to ubuntu-20.04 (#188)
As of Nov 9th, Github is experimenting with  transitioning `ubuntu-latest` label  to the Ubuntu 22.04 runner image. "During the transition, the label might refer to the runner image for either Ubuntu 20.04 or 22.04." which means if we get **lucky**, our `build_pypi` action will run on `ubuntu-22.04`-enabled runner and that will definitely fail `python3.6` sub-job since it is not prebuilt by GitHub therefore not available with Github action `setup-python` for the new ubuntu 22.04.

More information: https://github.blog/changelog/2022-11-09-github-actions-ubuntu-latest-workflows-will-use-ubuntu-22-04/

Ubuntu-18.04 is marked as deprecated and will be moved out of Github runner in the future, see https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners#supported-runners-and-hardware-resources. Moreover, the Python version is also slowly migrating towards 3.10 and deprecating 3.6, which PECOS future supported versions might be aware of and we will hold a discussion and PR to upgrade in the future.