Remove Ubuntu-18.04 in Actions
According to https://github.com/actions/runner-images/issues/6002, Ubuntu-18.04 is on a deprecation path and will periodically receive brownouts until 4/1/2023 to fully unsupported, we definitely need to remove that from the GitHub runner, and also bump the supported Ubuntu version of PECOS.

However, the newest Ubuntu-22.04 does not have a build-in Python 3.6 version which makes GitHub-hosted runner hard to run jobs, therefore not added here. But very soon, we need to make a decision to remove Python 3.6 and add Python 3.10 to supported versions of PECOS before next release and make change accordingly in CI/CD scripts.